Adds initial embedded wren parsing
Parsing a mixed Wren/HTML page, printing out all of the HTML as it's
written and passing on the Wren to be interpreted as code. Hopefully
escapes out everything that should be escaped so it runs correctly, but
at this point, who knows?

Currently allocates a stupidly large buffer for the output code - this
will be changed later to be a bit more solid.