Adds Web.getEnv() for list of headers
Returns the headers from the current request_rec through a table.

Wren's table API isn't committed yet, so this uses a diff that is a
small part of a pull request currently up:
  https://github.com/munificent/wren/pull/421

Now that we're manually applying a patch, we're going to be working
on top of a specific commit and rolling forward as the need arises.
Current Wren head for us:
  40c927f4402bb6ff74fe8aa257bf8042eeff6544

Returning from a foreign method with a map using more than 4-7 slots
causes problems unless another function is called afterwards. When
calling 'Web.getEnv()', if we don't call something like
'System.write("")' before trying to iterate it, the getEnv return value
is incomplete somehow (the error message says it doesn't implement its
iterator). This isn't related to the map API patch - lists suffer the
same problem.