Adds hacky wrapper methods returning lists
A current Wren bug makes it so that if a foreign method returns a list,
the returned list comes out as a 'Num' type until something else has
been run. As a (hopefully temporary) workaround, we wrap all foreign
methods that can return a list in a wrapper that performs an empty
System.write.