Cleans up leaked FILE* handle
Also playing around with allocations. Still not going to put the output
buffer in the request_rec pool without being sure how much resizing will
happen. File buffer will stay allocated just to make the function output
simpler (so we don't have to track whether or not to free it on the
other side).