Makes WrenVM remove modules after page load
Applies a patch to Wren that removes all loaded modules.

This is **bad** because it will reload all built-in modules for every
single page load. But it will also clear out all user-loaded modules,
which is what we want for modules to be reloaded each time.

It's not ideal but can be transparently fixed with a smarter system
later.